Move Etherpad service submenu actions into their own directory
